Develop

+++ server/src/app/modules/barcode/barcode.service.ts
@@ -1,4 +1,4 @@
-import { IBarCodeFilterRequest } from './barcode.interface';
+import { IBarCodeFilterRequest, IBarCodeStockRequest, ICreateBarCodeManually } from './barcode.interface';
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
@@ -11,6 +11,7 @@ import { IPaginationOptions } from '../../../interfaces/pagination';
 import { IGenericResponse } from '../../../interfaces/common';
 import { paginationHelpers } from '../../../helpers/paginationHelper';
 import { BarcodeRelationalFields, BarcodeRelationalFieldsMapper, BarcodeSearchableFields } from './barcode.constant';
+import { generateBarCode } from '../products/products.utils';
 
 // !----------------------------------get Single barcode ---------------------------------------->>>
 
@@ -83,7 +84,7 @@ const getProductBarcodeVarientWise = async (
   // Create a whereConditions object with AND conditions
   const whereConditions: Prisma.ProductVariationWhereInput = andConditions.length > 0 ? { AND: andConditions } : {};
 
-  // Retrieve product vriants with filtering and pagination
+  // Retrieve product variants with filtering and pagination
   const result = await prisma.productVariation.findMany({
     where: whereConditions,
     include: {
@@ -264,7 +265,7 @@ const getAllBarCodeForPrint = async (filters: IBarCodeFilterRequest, options: IP
   // Create a whereConditions object with AND conditions
   const whereConditions: Prisma.BarCodeWhereInput = andConditions.length > 0 ? { AND: andConditions } : {};
 
-  // Retrieve product vriants with filtering and pagination
+  // Retrieve product variants with filtering and pagination
   const result = await prisma.barCode.findMany({
     include: {
       variant: {
@@ -298,12 +299,120 @@ const getAllBarCodeForPrint = async (filters: IBarCodeFilterRequest, options: IP
   };
 };
 
-const getSingleVariant = async (variantId: string): Promise<any | null> => {
+// const getSingleVariant = async (variantId: string): Promise<any | null> => {
+//   if (!variantId) {
+//     throw new ApiError(httpStatus.BAD_REQUEST, 'variantId is required');
+//   }
+
+//   // Find the product using the productCode
+//   const result = await prisma.productVariation.findUnique({
+//     where: {
+//       variantId,
+//     },
+//     select: {
+//       variantId: true,
+//       productId: true,
+//       color: true,
+//       product: {
+//         select: {
+//           productName: true,
+//         },
+//       },
+//       _count: true,
+//       barCodes: {
+//         select: {
+//           barcodeId: true,
+//           code: true,
+//           barcodeStatus: true,
+//         },
+//       },
+//       // product: true,
+//     },
+//   });
+
+//   if (!result) {
+//     throw new ApiError(httpStatus.NOT_FOUND, 'Variation Not Found');
+//   }
+
+//   // console.log('product', result);
+
+//   return result;
+// };
+
+// Barcode Update
+
+const getSingleVariant = async (
+  variantId: string,
+  filters: IBarCodeFilterRequest,
+  options: IPaginationOptions
+): Promise<IGenericResponse<BarCode[]>> => {
+  // Calculate pagination options
+  const { limit, page, skip } = paginationHelpers.calculatePagination(options);
+
+  const { searchTerm, barcodeStatus, startDate, endDate, ...filterData } = filters;
+
+  // Define an array to hold filter conditions
+  const andConditions: Prisma.BarCodeWhereInput[] = [];
+
+  // Add search term condition if provided
+  if (searchTerm) {
+    andConditions.push({
+      OR: BarcodeSearchableFields.map((field: any) => ({
+        [field]: {
+          contains: searchTerm,
+          mode: 'insensitive',
+        },
+      })),
+    });
+  }
+
+  // Add filterData conditions if filterData is provided
+  if (Object.keys(filterData).length > 0) {
+    andConditions.push({
+      AND: Object.keys(filterData).map(key => {
+        if (BarcodeRelationalFields.includes(key)) {
+          return {
+            [BarcodeRelationalFieldsMapper[key]]: {
+              subCategoryName: (filterData as any)[key],
+            },
+          };
+        } else {
+          return {
+            [key]: {
+              equals: (filterData as any)[key],
+            },
+          };
+        }
+      }),
+    });
+  }
+
+  // Add barcodeStatus condition if provided
+  if (barcodeStatus) {
+    andConditions.push({
+      barcodeStatus: {
+        equals: barcodeStatus,
+      },
+    });
+  }
+
+  if (startDate && endDate) {
+    andConditions.push({
+      createdAt: {
+        gte: startDate, // Greater than or equal to startDate
+        lte: endDate, // Less than or equal to endDate
+      },
+    });
+  }
+
+  // This Code is not necessary
+  // const whereConditions: Prisma.BarCodeWhereInput = andConditions.length > 0 ? { AND: andConditions } : {};
+
   if (!variantId) {
     throw new ApiError(httpStatus.BAD_REQUEST, 'variantId is required');
   }
 
-  // Find the product using the productCode
+  // Find the product variation using the variantId
   const result = await prisma.productVariation.findUnique({
     where: {
       variantId,
@@ -317,15 +426,22 @@ const getSingleVariant = async (variantId: string): Promise<any | null> => {
           productName: true,
         },
       },
-      _count: true,
       barCodes: {
+        where: {
+          AND: [
+            { variantId }, // Assuming `variantId` is the correct field
+            ...andConditions,
+          ],
+        },
+        skip,
+        take: limit,
+        orderBy: options.sortBy && options.sortOrder ? { [options.sortBy]: options.sortOrder } : { updatedAt: 'desc' },
         select: {
           barcodeId: true,
           code: true,
           barcodeStatus: true,
         },
       },
-      // product: true,
     },
   });
 
@@ -333,12 +449,29 @@ const getSingleVariant = async (variantId: string): Promise<any | null> => {
     throw new ApiError(httpStatus.NOT_FOUND, 'Variation Not Found');
   }
 
-  // console.log('product', result);
+  // Count total matching barcodes for pagination
+  const total = await prisma.barCode.count({
+    where: {
+      AND: [
+        { variantId }, // Assuming `variantId` is the correct field
+        ...andConditions,
+      ],
+    },
+  });
 
-  return result;
+  // Return the paginated response
+  return {
+    meta: {
+      page,
+      limit,
+      total,
+      //@ts-ignore
+      totalPages: Math.ceil(total / limit),
+    },
+    //@ts-ignore
+    data: result.barCodes,
+  };
 };
-
-// Barcode Update
 
 const singleBarcodeUpdate = async (barcodeId: string, data: any): Promise<BarCode | null> => {
   console.log(data, 'data');
@@ -382,6 +515,10 @@ const deleteBarcode = async (barcodeId: string): Promise<BarCode | null> => {
     throw new ApiError(httpStatus.NOT_FOUND, 'Barcode Not Found');
   }
 
+  if (findBarcode.barcodeStatus !== 'INACTIVE') {
+    throw new ApiError(httpStatus.BAD_REQUEST, 'Only INACTIVE barcodes can be deleted');
+  }
+
   const result = await prisma.barCode.delete({
     where: {
       barcodeId,
@@ -392,8 +529,6 @@ const deleteBarcode = async (barcodeId: string): Promise<BarCode | null> => {
 };
 
 const deleteMultipleBarcode = async (barcodeIds: string[]): Promise<{ count: number }> => {
-  console.log('barcodeIds', barcodeIds);
-
   if (!barcodeIds || barcodeIds.length === 0) {
     throw new ApiError(httpStatus.BAD_REQUEST, 'barcodeIds are required');
   }
@@ -406,8 +541,16 @@ const deleteMultipleBarcode = async (barcodeIds: string[]): Promise<{ count: num
     },
   });
 
+  console.log(findBarcode, 'findBarcode');
+
   if (findBarcode.length === 0) {
     throw new ApiError(httpStatus.NOT_FOUND, 'Barcodes Not Found');
+  }
+
+  const invalidBarcodes = findBarcode.filter(barcode => barcode.barcodeStatus !== 'INACTIVE');
+
+  if (invalidBarcodes.length > 0) {
+    throw new ApiError(httpStatus.BAD_REQUEST, 'Only INACTIVE barcodes can be deleted');
   }
 
   const result = await prisma.barCode.deleteMany({
@@ -421,6 +564,62 @@ const deleteMultipleBarcode = async (barcodeIds: string[]): Promise<{ count: num
   return result;
 };
 
+//! stock increase
+const addBarCode = async (data: IBarCodeStockRequest): Promise<number> => {
+  const findVariant = await prisma.productVariation.findUnique({
+    where: {
+      variantId: data.variantId,
+    },
+  });
+
+  if (!findVariant) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Variant Not Found');
+  }
+
+  const codes = Array.from({ length: data.stock }, () => ({
+    code: generateBarCode(),
+    variantId: findVariant.variantId,
+  }));
+
+  const createdBarCodes = await prisma.barCode.createMany({ data: codes });
+
+  return createdBarCodes.count;
+};
+//! Manually QR Code Create
+const createQrCodeManually = async (data: ICreateBarCodeManually) => {
+  //
+  const findVariant = await prisma.productVariation.findUnique({
+    where: {
+      variantId: data?.variantId,
+    },
+  });
+
+  if (!findVariant) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Variant Not Found');
+  }
+  const findQrCode = await prisma.barCode.findUnique({
+    where: {
+      code: data?.code,
+    },
+  });
+
+  if (findQrCode) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Qr Code Already Exist');
+  }
+  // create
+  const newData = {
+    code: data.code,
+    variantId: findVariant.variantId,
+  };
+  const result = await prisma.barCode.create({
+    data: newData,
+  });
+  if (!result) {
+    throw new ApiError(httpStatus.BAD_REQUEST, 'QR Code creating failed');
+  }
+  return result;
+};
+
 export const BarcodeService = {
   getSingleBarCodeDetailsForKid,
   getProductBarcodeVarientWise,
@@ -430,4 +629,6 @@ export const BarcodeService = {
   singleBarcodeUpdate,
   deleteBarcode,
   deleteMultipleBarcode,
+  addBarCode,
+  createQrCodeManually,
 };
